Refactoring Layer struct neuralnet/layers module
