+++ delta/src/neuralnet/layers/conv_2d.rs
@@ -28,10 +28,10 @@
 //! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 use crate::activations::Activation;
-use crate::common::layer::Layer;
 use crate::common::shape::Shape;
 use crate::common::tensor_ops::Tensor;
 use crate::common::LayerError;
+use crate::neuralnet::layers::Layer;
 
 /// A struct representing a 2D convolutional layer.
 #[derive(Debug)]
