+++ delta/src/neuralnet/layers/mod.rs
@@ -33,8 +33,120 @@ pub mod max_pooling_2d;
 pub mod conv_2d;
 pub mod conv_1d;
 
+use std::fmt::Debug;
 pub use dense::Dense;
 pub use flatten::Flatten;
 pub use max_pooling_2d::MaxPooling2D;
 pub use conv_2d::Conv2D;
 pub use conv_1d::Conv1D;
+use crate::common::{LayerError, Shape, Tensor};
+use crate::optimizers::Optimizer;
+
+// A trait representing a neural network layer.
+pub trait Layer: Debug {
+    /// Builds the layer with the given input shape.
+    ///
+    /// # Arguments
+    ///
+    /// * `input_shape` - The shape of the input tensor.
+    fn build(&mut self, input_shape: Shape) -> Result<(), LayerError>;
+
+    /// Performs the forward pass of the layer.
+    ///
+    /// # Arguments
+    ///
+    /// * `input` - The input tensor.
+    ///
+    /// # Returns
+    ///
+    /// The output tensor after applying the layer.
+    fn forward(&mut self, input: &Tensor) -> Result<Tensor, LayerError>;
+
+    /// Performs the backward pass of the layer.
+    ///
+    /// # Arguments
+    ///
+    /// * `grad` - The gradient tensor from the next layer.
+    ///
+    /// # Returns
+    ///
+    /// The gradient tensor to be passed to the previous layer.
+    fn backward(&mut self, grad: &Tensor) -> Result<Tensor, LayerError>;
+
+    /// Returns the output shape of the layer.
+    ///
+    /// # Returns
+    ///
+    /// A `Shape` representing the output shape of the layer.
+    fn output_shape(&self) -> Result<Shape, LayerError>;
+
+    /// Returns the number of parameters in the layer.
+    ///
+    /// # Returns
+    ///
+    /// A tuple `(usize, usize)` representing the number of trainable and non-trainable parameters in the layer.
+    fn param_count(&self) -> Result<(usize, usize), LayerError>;
+
+    /// Returns the name of the layer.
+    ///
+    /// # Returns
+    ///
+    /// A `&str` representing the name of the layer.
+    fn name(&self) -> &str;
+
+    /// Returns the number of units in the layer.
+    ///
+    /// # Returns
+    ///
+    /// A `usize` representing the number of units in the layer. Default is 0.
+    fn units(&self) -> usize {
+        0
+    }
+
+    /// Updates the weights of the layer.
+    ///
+    /// # Arguments
+    ///
+    /// * `optimizer` - The optimizer to use.
+    fn update_weights(&mut self, optimizer: &mut Box<dyn Optimizer>) -> Result<(), LayerError>;
+
+    /// Returns the weights of the layer as a serializable format.
+    ///
+    /// # Returns
+    ///
+    /// A `serde_json::Value` containing the layer's weights.
+    fn get_weights(&self) -> serde_json::Value {
+        serde_json::json!({})
+    }
+
+    /// Returns the layer's configuration as a serializable format.
+    ///
+    /// # Returns
+    ///
+    /// A `serde_json::Value` containing the layer's configuration.
+    fn get_config(&self) -> serde_json::Value {
+        serde_json::json!({})
+    }
+
+    /// Returns the type name of the layer.
+    ///
+    /// # Returns
+    ///
+    /// A `String` representing the type name of the layer.
+    fn type_name(&self) -> String {
+        std::any::type_name::<Self>()
+            .split("::")
+            .last()
+            .unwrap_or("Unknown")
+            .to_string()
+    }
+}
+
+/// A struct representing the output of a layer.
+#[derive(Debug)]
+pub struct LayerOutput {
+    /// The output tensor of the layer.
+    pub output: Tensor,
+    /// The gradients tensor of the layer.
+    pub gradients: Tensor,
+}
